Facades: Allow empty charm config values
The following changes allow the charm config as a map and the charm
config as a yaml to work as intended. The code previously assumed that
there was always a yaml and that the charm config map would always
overwrite it.
The problem occurs when there isn't a charm yaml and it would blindly
overwrite values of the yaml. This caused empty strings to be converted
to nil, which then caused the --reset flow to be triggered. The --reset
flow essentially says, use the config defaults when you see nil, by
deleting any user config. That's something that we didn't want to
trigger. The change causes us to exit early if there isn't a yaml.

There is a much deeper problem and one that we will need to understand.
API versioning should essentially not touch the old code OR at the very
least have enough testing to ensure that any new changes keeps the old
behaviour. 100% code coverage in these scenarios would be beneficial to
ensure that we didn't trip up here. The problem with attempting to get
to that level, is the amount of levels you have to test to just get that
far. From API server all the way down to state, it's a large task and
maybe not worth it?